+++ packages/@expo/cli/src/start/server/metro/MetroBundlerDevServer.ts
@@ -128,13 +128,13 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     resources,
     template,
     devBundleUrl,
-    basePath,
+    baseUrl,
   }: {
     mode: 'development' | 'production';
     resources: SerialAsset[];
     template: string;
     /** asset prefix used for deploying to non-standard origins like GitHub pages. */
-    basePath: string;
+    baseUrl: string;
     devBundleUrl?: string;
   }): Promise<string> {
     if (!resources) {
@@ -144,7 +144,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     return htmlFromSerialAssets(resources, {
       dev: isDev,
       template,
-      basePath,
+      baseUrl,
       bundleUrl: isDev ? devBundleUrl : undefined,
     });
   }
@@ -279,11 +279,11 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     {
       mode,
       minify = mode !== 'development',
-      basePath,
+      baseUrl,
     }: {
       mode: 'development' | 'production';
       minify?: boolean;
-      basePath: string;
+      baseUrl: string;
     }
   ) {
     const devBundleUrlPathname = createBundleUrlPath({
@@ -319,7 +319,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
       resources,
       template: staticHtml,
       devBundleUrl: devBundleUrlPathname,
-      basePath,
+      baseUrl,
     });
     return {
       content,
@@ -448,7 +448,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
                 mode: options.mode ?? 'development',
                 minify: options.minify,
                 // No base path in development
-                basePath: '',
+                baseUrl: '',
               });
             },
           })
@@ -591,12 +591,12 @@ function htmlFromSerialAssets(
   {
     dev,
     template,
-    basePath,
+    baseUrl,
     bundleUrl,
   }: {
     dev: boolean;
     template: string;
-    basePath: string;
+    baseUrl: string;
     /** This is dev-only. */
     bundleUrl?: string;
   }
@@ -609,8 +609,8 @@ function htmlFromSerialAssets(
         return `<style data-expo-css-hmr="${metadata.hmrId}">` + source + '\n</style>';
       } else {
         return [
-          `<link rel="preload" href="${basePath}/${filename}" as="style">`,
-          `<link rel="stylesheet" href="${basePath}/${filename}">`,
+          `<link rel="preload" href="${baseUrl}/${filename}" as="style">`,
+          `<link rel="stylesheet" href="${baseUrl}/${filename}">`,
         ].join('');
       }
     })
@@ -622,7 +622,7 @@ function htmlFromSerialAssets(
     ? `<script src="${bundleUrl}" defer></script>`
     : jsAssets
         .map(({ filename }) => {
-          return `<script src="${basePath}/${filename}" defer></script>`;
+          return `<script src="${baseUrl}/${filename}" defer></script>`;
         })
         .join('');