+++ packages/babel-preset-expo/src/index.ts
@@ -1,9 +1,16 @@
 import { ConfigAPI, PluginItem, TransformOptions } from '@babel/core';
 
-import { getBundler, getInlineEnvVarsEnabled, getIsDev, getIsProd, hasModule } from './common';
+import {
+  getBaseUrl,
+  getBundler,
+  getInlineEnvVarsEnabled,
+  getIsDev,
+  getIsProd,
+  hasModule,
+} from './common';
 import { expoInlineManifestPlugin } from './expo-inline-manifest-plugin';
 import { expoRouterBabelPlugin } from './expo-router-plugin';
-import { expoInlineEnvVars } from './inline-env-vars';
+import { expoInlineEnvVars, expoInlineTransformEnvVars } from './inline-env-vars';
 import { lazyImports } from './lazyImports';
 
 type BabelPresetExpoPlatformOptions = {
@@ -53,6 +60,7 @@ function babelPresetExpo(api: ConfigAPI, options: BabelPresetExpoOptions = {}): 
   let platform = api.caller((caller) => (caller as any)?.platform);
   const engine = api.caller((caller) => (caller as any)?.engine) ?? 'default';
   const isDev = api.caller(getIsDev);
+  const baseUrl = api.caller(getBaseUrl);
   // Unlike `isDev`, this will be `true` when the bundler is explicitly set to `production`,
   // i.e. `false` when testing, development, or used with a bundler that doesn't specify the correct inputs.
   const isProduction = api.caller(getIsProd);
@@ -126,6 +134,15 @@ function babelPresetExpo(api: ConfigAPI, options: BabelPresetExpoOptions = {}): 
   if (aliasPlugin) {
     extraPlugins.push(aliasPlugin);
   }
+
+  extraPlugins.push([
+    expoInlineTransformEnvVars,
+    {
+      // These values should not be prefixed with `EXPO_PUBLIC_`, so we don't
+      // squat user-defined environment variables.
+      EXPO_BASE_URL: baseUrl,
+    },
+  ]);
 
   // Only apply in non-server, for metro-only, in production environments, when the user hasn't disabled the feature.
   // Webpack uses DefinePlugin for environment variables.