+++ packages/expo-router/src/fork/getStateFromPath.ts
@@ -45,7 +45,7 @@ type ParsedRoute = {
 
 export function getUrlWithReactNavigationConcessions(
   path: string,
-  basePath: string | undefined = Constants.expoConfig?.experiments?.basePath
+  baseUrl: string | undefined = Constants.expoConfig?.experiments?.baseUrl
 ) {
   let parsed: URL;
   try {
@@ -64,10 +64,10 @@ export function getUrlWithReactNavigationConcessions(
   return {
     // The slashes are at the end, not the beginning
     nonstandardPathname:
-      stripBasePath(pathname, basePath).replace(/^\/+/g, '').replace(/\/+$/g, '') + '/',
+      stripBaseUrl(pathname, baseUrl).replace(/^\/+/g, '').replace(/\/+$/g, '') + '/',
 
     // React Navigation doesn't support hashes, so here
-    inputPathnameWithoutHash: stripBasePath(path, basePath).replace(/#.*$/, ''),
+    inputPathnameWithoutHash: stripBaseUrl(path, baseUrl).replace(/#.*$/, ''),
   };
 }
 
@@ -778,24 +778,24 @@ const parseQueryParams = (path: string, parseConfig?: Record<string, (value: str
   return Object.keys(params).length ? params : undefined;
 };
 
-const basePathCache = new Map<string, RegExp>();
+const baseUrlCache = new Map<string, RegExp>();
 
-function getBasePathRegex(basePath: string) {
-  if (basePathCache.has(basePath)) {
-    return basePathCache.get(basePath)!;
+function getBaseUrlRegex(baseUrl: string) {
+  if (baseUrlCache.has(baseUrl)) {
+    return baseUrlCache.get(baseUrl)!;
   }
-  const regex = new RegExp(`^\\/?${escape(basePath)}`, 'g');
-  basePathCache.set(basePath, regex);
+  const regex = new RegExp(`^\\/?${escape(baseUrl)}`, 'g');
+  baseUrlCache.set(baseUrl, regex);
   return regex;
 }
 
-export function stripBasePath(
+export function stripBaseUrl(
   path: string,
-  basePath: string | undefined = Constants.expoConfig?.experiments?.basePath
+  baseUrl: string | undefined = Constants.expoConfig?.experiments?.baseUrl
 ) {
   if (process.env.NODE_ENV !== 'development') {
-    if (basePath) {
-      const reg = getBasePathRegex(basePath);
+    if (baseUrl) {
+      const reg = getBaseUrlRegex(baseUrl);
       return path.replace(/^\/+/g, '/').replace(reg, '');
     }
   }