+++ packages/@expo/cli/CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### 🎉 New features
 
+- Rename `basePath` to `baseUrl` and expose to bundles with `process.env.EXPO_BASE_URL`.
 - Move environment variables production inlining to `babel-preset-expo` to support source maps. ([#25239](https://github.com/expo/expo/pull/25239) by [@EvanBacon](https://github.com/EvanBacon))
 - Add support for chaining the Metro resolver locally. ([#25148](https://github.com/expo/expo/pull/25148) by [@EvanBacon](https://github.com/EvanBacon))
 - Pass engine to transformer for production bundling. ([#25175](https://github.com/expo/expo/pull/25175) by [@EvanBacon](https://github.com/EvanBacon))